refactor: create file and directory to centralize domain enums
