+++ src/common/enums/domain.enums.ts
@@ -1,0 +1,25 @@
+import {
+  FrequencyStatus,
+  Race,
+  Gender,
+  FormType,
+  UserStatus,
+  StudentStatus,
+  SocialProgram,
+  EmploymentStatus,
+  SchoolYear,
+  EducationLevel
+} from '@prisma/client';
+
+export {
+  FrequencyStatus,
+  Race,
+  Gender,
+  FormType,
+  UserStatus,
+  StudentStatus,
+  SocialProgram,
+  EmploymentStatus,
+  SchoolYear,
+  EducationLevel
+};
